requirements.txt
pandas==1.3.5
scikit-learn==1.2.0
joblib==1.2.0
google-cloud-storage==2.10.0
